AP7E-3426 Unable to play video files from Google Drive when using Chromium
Why:
Users can not play video files from Google Drive when using Chromium.

How:
We found that when setting Chromium to mobile mode,
we can play videos on Google Drive without issues,
whereas it doesn't work in desktop mode.

We also found that Chromium can be the default in desktop mode,
with exceptions for certain web pages. In other words,
it can be configured to use mobile mode only when browsing the web version of Google Drive
(allowing users to watch videos normally), while it remains in desktop mode for other web pages.

Considering the complexity of setting VLC as the default player,
we decided to modify Chromium source codes to add the desktop mode exception for Google Drive as the default setting in Chromium.